requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==2.2.0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==4.0.3
attrs==20.3.0
audioread==3.0.1
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
cachetools==5.5.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.3.2
cmake==3.30.5
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
decorator==4.4.2
decord==0.6.0
deepspeed==0.12.6
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.8
einops==0.8.0
entrypoints==0.3
et-xmlfile==2.0.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.24.5
idna==2.10
imageio==2.35.0
imageio-ffmpeg==0.5.1
importlib-metadata==8.2.0
importlib-resources==6.4.0
iopath==0.1.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
jinja2==3.1.4
joblib==1.4.2
json5==0.9.5
jsonlines==4.0.0
jsonschema==3.2.0
jupyter-client==6.2.0
jupyter-core==4.7.1
jupyter-packaging==0.9.1
jupyter-server==1.6.1
jupyterlab==3.0.14
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
kiwisolver==1.4.5
lazy-loader==0.4
librosa==0.10.2.post1
lit==18.1.8
llvmlite==0.43.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
mistune==0.8.4
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nbclassic==0.2.7
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==3.2.1
ninja==1.11.1.1
notebook==6.3.0
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.555.43
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==20.9
pandas==2.2.2
pandocfilters==1.4.3
parso==0.8.2
peft==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
pip==21.0.1
platformdirs==4.2.2
pooch==1.8.2
portalocker==2.10.1
proglog==0.1.10
prometheus-client==0.10.1
prompt-toolkit==3.0.18
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.0.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.20
pydantic==2.8.2
pydantic-core==2.20.1
pydub==0.25.1
Pygments==2.8.1
PyJWT==2.8.0
pynvml==11.5.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.9.0.post0
pytz==2021.1
PyYAML==6.0.1
pyzmq==22.0.3
regex==2024.7.24
requests==2.32.3
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
Send2Trash==1.5.0
sentencepiece==0.2.0
setuptools==75.1.0
six==1.15.0
sniffio==1.2.0
soundfile==0.12.1
soxr==0.4.0
sympy==1.13.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
terminado==0.9.4
testpath==0.4.4
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.15.2
tomlkit==0.7.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.1
tqdm==4.66.4
traitlets==5.0.5
transformers==4.37.2
transformers-stream-generator==0.0.5
triton==2.0.0
typing-extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.4
wcwidth==0.2.13
webencodings==0.5.1
werkzeug==3.0.1
wheel==0.36.2
xxhash==3.4.1
yarl==1.9.4
zhipuai==2.1.5.20241204
zipp==3.19.2
